sign in and register now return a token, this token now authenticates on subsequent log-ins and boots the user from the splash page straight to the maps activity.

+++ app/src/main/java/com/camvy/kai/coolmaps/Networking/UserState.java
@@ -17,8 +17,8 @@ public class UserState {
     public static Badge getBadge(Context context){
         SharedPreferences pref = PreferenceManager.getDefaultSharedPreferences(context);
 
-        if (pref.getInt("user_id", 0) != 0){
-            userBadge = new Badge(pref.getInt("user_id", 0), pref.getString("session_token", null));
+        if (pref.getFloat("user_id", 0) != 0){
+            userBadge = new Badge(pref.getFloat("user_id", 0), pref.getString("session_token", null));
         }
 
         return userBadge;
@@ -27,7 +27,7 @@ public class UserState {
     public static void setBadge(Badge badgeTobeSaved, Context context){
         SharedPreferences pref = PreferenceManager.getDefaultSharedPreferences(context);
         SharedPreferences.Editor e = pref.edit();
-        e.putInt("user_id", badgeTobeSaved.getUser_id());
+        e.putFloat("user_id", badgeTobeSaved.getUser_id());
         e.putString("session_token", badgeTobeSaved.getSession_token());
         e.commit();
     }
